pyproject.toml
[tool.poetry]
name = "fishywatch"
version = "0.1.0"
description = ""
authors = ["odysa <piensengxv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
tldextract = "^5.0.1"
lxml = "^4.9.3"
aiohttp = "^3.8.6"
aiofiles = "^23.2.1"
asyncpg = "^0.29.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-sugar = "^0.9.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-select = ["I"]